Security incident Report
This is an example of how I wrote an incident report in the aftermath of brute force compromise of a DNS server. In this hypothetical situation, I documented the analysis of the incident, user experience, containment and remediation actions, and business continuity measures I took following the attack. I concluded with recommendations on how to forestall similar future attacks.